+++ activerecord/CHANGELOG.md
@@ -1,5 +1,18 @@
+*   Fix interactions between `:before_add` callbacks and nested attributes
+    assignment of `has_many` associations, when the association was not
+    yet loaded:
+
+    - A `:before_add` callback was being called when a nested attributes
+      assignment assigned to an existing record.
+
+    - Nested Attributes assignment did not affect the record in the
+      association target when a `:before_add` callback triggered the
+      loading of the association
+
+    *Jörg Schray*
+
 *   Allow enable_extension migration method to be revertible.
- 
+
     *Eric Tipton*
 
 *   Type cast hstore values on write, so that the value is consistent